feat: Add return values to setAddressArray and setBytesArray functions
This commit modifies the setAddressArray and setBytesArray functions in the HyperAGI_Storage contract to return the length of the corresponding array after the value is set. This change provides useful feedback to the caller and improves the usability of these functions.

Note: This commit message follows the conventional format of "type: description" and provides a clear and concise summary of the code changes.