feat: Upgrade HyperAGI_Agent_Token contract and deploy proxy instance
This commit upgrades the HyperAGI_Agent_Token contract by using the `upgrades` library from Hardhat. It also deploys a proxy instance of the contract with the specified parameters. This change improves the functionality and ensures the use of the latest version of the contract.

Note: This commit message follows the conventional format of "type: description" and provides a clear and concise summary of the code changes.